+++ lib/ui/profile/ProfileInfoEditPage.dart
@@ -1,7 +1,9 @@
 
 import 'dart:async';
+import 'dart:math';
 import 'dart:typed_data';
 
+import 'package:dropdown_button2/dropdown_button2.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:neom/ext/Auth2.dart';
@@ -78,8 +80,17 @@ class ProfileInfoEditPageState extends ProfileDirectoryMyInfoBasePageState<Profi
   AudioPlayer? _audioPlayer;
 
   double _screenInsetsBottom = 0;
+  double? _visibilityDropdownItemsWidth;
   Timer? _onScreenInsetsBottomChangedTimer;
 
+  static const double _buttonIconSize = 16;
+  static const double _dropdownItemInnerIconPaddingX = 6;
+  static const double _dropdownButtonInnerIconPaddingX = 12;
+  static const double _dropdownButtonChevronIconSize = 10;
+  static const EdgeInsetsGeometry _dropdownMenuItemPadding = const EdgeInsets.symmetric(horizontal: 16, vertical: 16);
+  static const EdgeInsetsGeometry _dropdownButtonPadding = const EdgeInsets.only(left: 16, right: 8, top: 15, bottom: 15);
+
+  bool _isFieldAvailable(_ProfileField field) => (fieldAvailabilities[widget.authType?.loginType]?.contains(field) == true);
   bool get _showProfileCommands => (widget.onboarding == false);
   bool get _showPrivacyControls => (widget.onboarding == false) && FlexUI().isPrivacyAvailable;
   bool get _showNameControls => !_hasProfileName;
@@ -162,6 +173,13 @@ class ProfileInfoEditPageState extends ProfileDirectoryMyInfoBasePageState<Profi
         setStateIfMounted(() {});
       });
     }
+
+    double visibilityDropdownItemsWidth = _evaluateVisibilityDropdownItemsWidth();
+    if (_visibilityDropdownItemsWidth != visibilityDropdownItemsWidth) {
+      setStateIfMounted(() {
+        _visibilityDropdownItemsWidth = visibilityDropdownItemsWidth;
+      });
+    }
   }
 
   @override
@@ -209,7 +227,7 @@ class ProfileInfoEditPageState extends ProfileDirectoryMyInfoBasePageState<Profi
     Map<String, String>? get _photoAuthHeaders => DirectoryProfilePhotoUtils.authHeaders;
 
     Widget get _photoWidget => Stack(children: [
-      Padding(padding: EdgeInsets.only(left: 8, right: 8, bottom: 20), child:
+      Padding(padding: EdgeInsets.only(left: 16, right: 16, bottom: 20), child:
         DirectoryProfilePhoto(
           key: _photoKey,
           photoUrl: _photoImageUrl,
@@ -220,13 +238,15 @@ class ProfileInfoEditPageState extends ProfileDirectoryMyInfoBasePageState<Profi
       ),
       Positioned.fill(child:
         Align(alignment: _showPrivacyControls ? Alignment.bottomLeft : Alignment.bottomRight, child:
-          _editPhotoButton
+          Padding(padding: EdgeInsets.symmetric(horizontal: 8), child:
+            _editPhotoButton
+          )
         )
       ),
       if (_showPrivacyControls)
         Positioned.fill(child:
           Align(alignment: Alignment.bottomRight, child:
-            _togglePhotoVisibilityButton
+            _editPhotoVisibilityButton
           )
         )
     ],);
@@ -321,10 +341,20 @@ class ProfileInfoEditPageState extends ProfileDirectoryMyInfoBasePageState<Profi
       Analytics().logAlert(text: _clearPhotoPrompt(language: 'en'), selection: 'OK');
     }
 
-    Widget get _togglePhotoVisibilityButton =>
-      _photoIconButton(_visibilityIcon(_fieldVisibilities[_ProfileField.photoUrl], _fieldTextNotEmpty[_ProfileField.photoUrl] == true),
-        onTap: (_fieldTextNotEmpty[_ProfileField.photoUrl] == true) ? () => _onToggleFieldVisibility(profileField: _ProfileField.photoUrl) : null,
+    Widget get _editPhotoVisibilityButton =>
+      _photoVisibilityDropdown; // _photoVisibilityToggleButton
+
+    Widget get _photoVisibilityDropdown =>
+      _visibilityDropdown(_ProfileField.photoUrl,
+        buttonPadding: EdgeInsets.only(left: 8, right: 6, top: 10, bottom: 10),
+        buttonInnerIconPadding: 8
       );
+
+    // ignore: unused_element
+    Widget get _photoVisibilityToggleButton =>
+    _photoIconButton(_visibilityIcon(_ProfileField.photoUrl),
+      onTap: (_fieldTextNotEmpty[_ProfileField.photoUrl] == true) ? () => _onToggleFieldVisibility(_ProfileField.photoUrl) : null,
+    );
 
     Widget _photoIconButton(Widget? icon, { void Function()? onTap, bool progress = false}) =>
       InkWell(onTap: onTap, child:
@@ -343,8 +373,6 @@ class ProfileInfoEditPageState extends ProfileDirectoryMyInfoBasePageState<Profi
           ),
         )
       );
-
-    static const double _buttonIconSize = 16;
 
     Widget get _staticNameWidget =>
       Text(_staicNameText ?? '', style: nameTextStyle, textAlign: TextAlign.center,);
@@ -708,11 +736,15 @@ class ProfileInfoEditPageState extends ProfileDirectoryMyInfoBasePageState<Profi
   Widget _visibilityButton({ _ProfileField? profileField, Auth2PublicAccountIdentifier? identifier, bool locked = false}) {
     Auth2FieldVisibility? visibility = _getFieldVisibility(profileField: profileField, identifier: identifier);
     bool textNotEmpty = profileField != null ? (_fieldTextNotEmpty[profileField] == true) : (identifier != null ? _identifierTextNotEmpty[identifier] == true : false);
-    return _iconButton(
-      icon: _visibilityIcon(visibility, textNotEmpty, locked: locked),
-      onTap: (textNotEmpty && !locked) ? () => _onToggleFieldVisibility(profileField: profileField, identifier: identifier) : null,
-    );
+    return _visibilityDropdown(visibility, textNotEmpty, locked: locked); // _visibilityToggleButton(field, locked: locked);
   }
+
+  // ignore: unused_element
+  Widget _visibilityToggleButton(_ProfileField field, { bool locked = false }) =>
+    _iconButton(
+      icon: _visibilityIcon(field, locked: locked),
+      onTap: ((_fieldTextNotEmpty[field] == true) && !locked) ? () => _onToggleFieldVisibility(field) : null,
+    );
 
   Widget _iconButton({ Widget? icon, void Function()? onTap, bool progress = false}) =>
     InkWell(onTap: onTap, child:
@@ -735,6 +767,16 @@ class ProfileInfoEditPageState extends ProfileDirectoryMyInfoBasePageState<Profi
     }
   }
 
+  Widget? _visibilityDropdownIcon(Auth2FieldVisibility? visibility, bool textNotEmpty, { bool locked = false} ) {
+    if (locked) {
+      return _lockIcon;
+    } else if (_permittedVisibility.contains(visibility) && textNotEmpty) {
+      return _publicDropdownIcon;
+    } else {
+      return _privateDropdownIcon;
+    }
+  }
+
   Auth2FieldVisibility? _getFieldVisibility({_ProfileField? profileField, Auth2PublicAccountIdentifier? identifier}) {
     if (profileField != null) {
       return _fieldVisibilities[profileField];
@@ -744,16 +786,162 @@ class ProfileInfoEditPageState extends ProfileDirectoryMyInfoBasePageState<Profi
     return null;
   }
 
-  Widget? get _editIcon => Styles().images.getImage('edit', color: Styles().colors.fillColorPrimary, size: _buttonIconSize);
-  Widget? get _trashIcon => Styles().images.getImage('trash', color: Styles().colors.fillColorPrimary, size: _buttonIconSize);
-  Widget? get _publicIcon => Styles().images.getImage('eye', color: Styles().colors.fillColorSecondary, size: _buttonIconSize);
-  Widget? get _privateIcon => Styles().images.getImage('eye-slash', color: Styles().colors.mediumGray2, size: _buttonIconSize);
-  Widget? get _lockIcon => Styles().images.getImage('lock', color: Styles().colors.fillColorSecondary, size: _buttonIconSize);
-  Widget? get _playIcon => Styles().images.getImage('play', color: Styles().colors.fillColorPrimary, size: _buttonIconSize);
-  Widget? get _pauseIcon => Styles().images.getImage('pause', color: Styles().colors.fillColorPrimary, size: _buttonIconSize);
-  //Widget? get _stopIcon => Styles().images.getImage('stop', color: Styles().colors.fillColorPrimary, size: _editButtonIconSize);
+  static Widget? get _editIcon => Styles().images.getImage('edit', color: Styles().colors.fillColorPrimary, size: _buttonIconSize);
+  static Widget? get _trashIcon => Styles().images.getImage('trash', color: Styles().colors.fillColorPrimary, size: _buttonIconSize);
+  static Widget? get _publicIcon => Styles().images.getImage('eye', color: Styles().colors.fillColorSecondary, size: _buttonIconSize);
+  static Widget? get _privateIcon => Styles().images.getImage('eye-slash', color: Styles().colors.mediumGray2, size: _buttonIconSize);
+  static Widget? get _lockIcon => Styles().images.getImage('lock', color: Styles().colors.fillColorSecondary, size: _buttonIconSize);
+  static Widget? get _playIcon => Styles().images.getImage('play', color: Styles().colors.fillColorPrimary, size: _buttonIconSize);
+  static Widget? get _pauseIcon => Styles().images.getImage('pause', color: Styles().colors.fillColorPrimary, size: _buttonIconSize);
+  static Widget? get _publicDropdownIcon => Styles().images.getImage('earth-americas', color: Styles().colors.fillColorSecondary, size: _buttonIconSize);
+  static Widget? get _privateDropdownIcon => Styles().images.getImage('earth-americas', color: Styles().colors.mediumGray2, size: _buttonIconSize);
+  static Widget? get _lockDropdownIcon => Styles().images.getImage('lock', color: Styles().colors.mediumGray2, size: _buttonIconSize);
+  static Widget? get _chevronDropdownIcon => Styles().images.getImage('chevron-down', color: Styles().colors.mediumGray2, size: _dropdownButtonChevronIconSize);
+  static Widget? get _redioOnDropdownIcon => Styles().images.getImage('radio-button-on', size: _buttonIconSize);
+  static Widget? get _redioOffDropdownIcon => Styles().images.getImage('radio-button-off', size: _buttonIconSize);
 
-  Decoration get _controlDecoration => BoxDecoration(
+  //static  Widget? get _stopIcon => Styles().images.getImage('stop', color: Styles().colors.fillColorPrimary, size: _editButtonIconSize);
+
+  Widget _visibilityDropdown(_ProfileField field, {
+    bool locked = false,
+    EdgeInsetsGeometry buttonPadding = _dropdownButtonPadding,
+    double buttonInnerIconPadding = _dropdownButtonInnerIconPaddingX,
+  }) =>
+      DropdownButtonHideUnderline(child:
+      DropdownButton2<Auth2FieldVisibility>(
+        dropdownStyleData: DropdownStyleData(
+          width: _visibilityDropdownItemsWidth ??= _evaluateVisibilityDropdownItemsWidth(),
+          direction: DropdownDirection.left,
+          decoration: _controlDecoration,
+        ),
+        customButton: locked ? _visibilityDropdownLockedButton : _visibilityDropdownButton(field,
+          padding: buttonPadding,
+          innerIconPadding: buttonInnerIconPadding,
+        ),
+        isExpanded: false,
+        items: _visibilityDropdownItems(field),
+        onChanged: ((_fieldTextNotEmpty[field] == true) && !locked) ? (Auth2FieldVisibility? visibility) => _onDropdownFieldVisibility(field, visibility) : null,
+      ),
+      );
+
+  Widget _visibilityDropdownButton(_ProfileField field, {
+    bool locked = false,
+    EdgeInsetsGeometry padding = _dropdownButtonPadding,
+    double innerIconPadding = _dropdownButtonInnerIconPaddingX,
+  }) =>
+      Container(decoration: _controlDecoration, child:
+      Padding(padding: padding, child:
+      Row(mainAxisSize: MainAxisSize.min, children: [
+        SizedBox(width: _buttonIconSize, height: _buttonIconSize, child:
+        Center(child: _visibilityDropdownIcon(field, locked: locked),)
+        ),
+        Padding(padding: EdgeInsets.only(left: innerIconPadding), child:
+        SizedBox(width: _dropdownButtonChevronIconSize, height: _dropdownButtonChevronIconSize, child:
+        Center(child:
+        locked ? null : _chevronDropdownIcon,
+        )
+        )
+        )
+      ],)
+      )
+      );
+
+  Widget get _visibilityDropdownLockedButton =>
+      Container(decoration: _controlDecoration, child:
+      Padding(padding: EdgeInsets.only(left: 23, right: 23, top: 15, bottom: 15), child:
+      SizedBox(width: _buttonIconSize, height: _buttonIconSize, child:
+      Center(child: _lockIcon,)
+      ),
+      )
+      );
+
+  List<DropdownMenuItem<Auth2FieldVisibility>> _visibilityDropdownItems(_ProfileField field) {
+    List<DropdownMenuItem<Auth2FieldVisibility>> items = <DropdownMenuItem<Auth2FieldVisibility>>[];
+    Auth2FieldVisibility selectedFieldVisibility = profileFieldVisibility(field);
+    for (Auth2FieldVisibility fieldVisibility in Auth2FieldVisibility.values.reversed) {
+      if ((fieldVisibility == Auth2FieldVisibility.private) || _permittedVisibility.contains(fieldVisibility)) {
+        items.add(_visibilityDropdownItem(fieldVisibility, selected: selectedFieldVisibility == fieldVisibility));
+      }
+    }
+    return items;
+  }
+
+  DropdownMenuItem<Auth2FieldVisibility> _visibilityDropdownItem(Auth2FieldVisibility visibility, { bool selected = false}) =>
+      DropdownMenuItem<Auth2FieldVisibility>(
+        value: visibility,
+        child: Semantics(label: visibility.semanticLabel, container: true, button: true, child:
+        Column(mainAxisSize: MainAxisSize.min, crossAxisAlignment: CrossAxisAlignment.start, children: [
+          Row(mainAxisSize: MainAxisSize.max, children: [
+            Padding(padding: EdgeInsets.only(right: _dropdownItemInnerIconPaddingX), child:
+            SizedBox(width: _buttonIconSize, height: _buttonIconSize, child:
+            Center(child: visibility.displayDropdownItemIcon)
+            )
+            ),
+            Expanded(child:
+            Text(visibility.displayTitle,
+              overflow: TextOverflow.ellipsis,
+              style: selected ? _selectedDropdownItemTextStyle : _regularDropdownItemTextStyle,
+              semanticsLabel: "",
+            ),
+            ),
+            Padding(padding: EdgeInsets.only(left: _dropdownItemInnerIconPaddingX), child:
+            SizedBox(width: _buttonIconSize, height: _buttonIconSize, child:
+            Center(child: selected ? _redioOnDropdownIcon : _redioOffDropdownIcon)
+            )
+            )
+          ],),
+          if (visibility.displayDescription.isNotEmpty)
+            Padding(padding: EdgeInsets.symmetric(horizontal: _dropdownItemInnerIconPaddingX + _buttonIconSize), child:
+            Text(visibility.displayDescription,
+              overflow: TextOverflow.ellipsis,
+              style: _descriptionDropdownItemTextStyle,
+              semanticsLabel: "",
+            ),
+            )
+        ],)
+        ),
+      );
+
+  double _evaluateVisibilityDropdownItemsWidth() {
+    double maxTextWidth = 0;
+    for (Auth2FieldVisibility fieldVisibility in Auth2FieldVisibility.values) {
+      if ((fieldVisibility == Auth2FieldVisibility.private) || _permittedVisibility.contains(fieldVisibility)) {
+        final Size textSizeFull = (TextPainter(
+          text: TextSpan(text: fieldVisibility.displayTitle, style: _selectedDropdownItemTextStyle,),
+          textScaler: MediaQuery.of(context).textScaler,
+          textDirection: TextDirection.ltr,
+        )..layout()).size;
+        if (maxTextWidth < textSizeFull.width) {
+          maxTextWidth = textSizeFull.width;
+        }
+        final Size descriptionSizeFull = (TextPainter(
+          text: TextSpan(text: fieldVisibility.displayDescription, style: _selectedDropdownItemTextStyle,),
+          textScaler: MediaQuery.of(context).textScaler,
+          textDirection: TextDirection.ltr,
+        )..layout()).size;
+        if (maxTextWidth < descriptionSizeFull.width) {
+          maxTextWidth = descriptionSizeFull.width;
+        }
+      }
+    }
+    double dropdownItemWidth = (maxTextWidth * 5 / 3) + 2 * (_buttonIconSize + _dropdownItemInnerIconPaddingX) + _dropdownMenuItemPadding.horizontal;
+    return min(dropdownItemWidth, MediaQuery.of(context).size.width * 2 / 3);
+  }
+
+
+  TextStyle? get _selectedDropdownItemTextStyle => Styles().textStyles.getTextStyle("widget.item.regular.extra_fat");
+  TextStyle? get _regularDropdownItemTextStyle => Styles().textStyles.getTextStyle("widget.item.regular.semi_fat");
+  TextStyle? get _descriptionDropdownItemTextStyle => Styles().textStyles.getTextStyle("widget.item.small.semi_fat");
+
+  void _onDropdownFieldVisibility(_ProfileField field, Auth2FieldVisibility? visibility) {
+    Analytics().logSelect(target: 'Select $field Visibility $visibility');
+    setState(() {
+      _fieldVisibilities[field] = visibility;
+    });
+  }
+
+
+  BoxDecoration get _controlDecoration => BoxDecoration(
     color: Styles().colors.surface,
     border: Border.all(color: Styles().colors.surfaceAccent, width: 1),
     borderRadius: BorderRadius.all(Radius.circular(8)),
@@ -1188,3 +1376,35 @@ extension _Auth2UserProfileFieldsVisibilityUtils on Auth2UserProfileFieldsVisibi
     );
 }
 
+///////////////////////////////////////////
+// _Auth2FieldVisibilityUI
+
+extension _Auth2FieldVisibilityUI on Auth2FieldVisibility {
+
+  String get displayTitle {
+    switch(this) {
+      case Auth2FieldVisibility.public: return Localization().getStringEx('panel.profile.info.directory_visibility.dropdown.public.title', 'Public');
+      case Auth2FieldVisibility.connections: return Localization().getStringEx('panel.profile.info.directory_visibility.dropdown.connections.title', 'Only My Connections');
+      case Auth2FieldVisibility.private: return Localization().getStringEx('panel.profile.info.directory_visibility.dropdown.private.title', 'Only me');
+    }
+  }
+
+  String get displayDescription {
+    switch(this) {
+      case Auth2FieldVisibility.public: return Localization().getStringEx('panel.profile.info.directory_visibility.dropdown.public.description', 'Anyone can view');
+      case Auth2FieldVisibility.connections: return Localization().getStringEx('panel.profile.info.directory_visibility.dropdown.connections.description', '');
+      case Auth2FieldVisibility.private: return Localization().getStringEx('panel.profile.info.directory_visibility.dropdown.private.description', '');
+    }
+  }
+
+  String get semanticLabel =>
+    "$displayTitle $displayDescription";
+
+  Widget? get displayDropdownItemIcon {
+    switch(this) {
+      case Auth2FieldVisibility.public: return ProfileInfoEditPageState._publicDropdownIcon;
+      case Auth2FieldVisibility.connections: return ProfileInfoEditPageState._lockDropdownIcon;
+      case Auth2FieldVisibility.private: return ProfileInfoEditPageState._lockDropdownIcon;
+    }
+  }
+}